Added files for new functions.
These files add a place for the parser, symbol table code, and the code
module.

Two structures for the code module have been created named tri and sept.

We are not using bitwise manipulation for these structures to make the code
easier to understand and maintain code portability between different
architectures.  Moreover, the datatype written to file is type char so this
happens to fit well with the outcome desired.

Changed README to reflect location of blessed repository.

There are currently two repos and I don't want anyone to get confused
or stuck with old code.