Moved functions in main into parser and error files.
Have finished moving functions in main to correct files.  This should
eliminate much of the issues with not following the API mentioned int the book.

Have created error files to store error and warning message code.
This functionality is not requred but have been added to increase
the usefulness of the system when completed.