Implemented comp function and updated command_type.
Command type was not working for all command versions.
This version works a little better but still needs more work.

Implemented comp function and made changes to dest.
Function dest now verifies some of the syntax of the code.
Function comp also verifies delimiters in the assembly and
exits if some thing is wrong.

Printing dest and comp in main now checks for NULL pointers
that are now possible.