Updated parser functions.
Function dest now works will all types of C Commands.
Version now checks for '=' sign before parsing.

Function comp is now more intelligent.  Verifies input
command for presence of ';' or '='.  Throws error 9
when token not found as this should not happen.

Set optimization flags in Makefile to 0.  GCC was
generating binary code that was broken.  This might be set
back at a later time when I find a fix.