Fixed bug in enc_dest and enc_jump.
Previous version was suffering from a bug in object code.
Value of pointer mnemonic was changing unexpectedly.
This was tested on two systems:

Mandriva 2009Centos 5.4
gcc4.3.24.1.2
gdb6.8.26.8.37
make3.813.81
glibc2.8.12.5-42

Current version copies mnemonic data into separate array
and with pointer.  This is the long way around this bug.

In the process of making these changes, I have added the
pendantic switch to the Makefile.