Added symbol hash code.
This will be used to hash symbols in assembly code.
Named loops and variables will be available to the assembly
programmer.  This makes the language less difficult to understand.

Hash code in this version is known to work but has not tied to the
rest of the program.

Made some changes to ERRORS to make documentation more readable.
Error number 13 is now present. This error occurs when symbol table
is full and cannot hold any more values.  If this error should ever
occur, program should be recompiled with a higher primary number
for HASH_SIZE.